Refactored ScannerController, moving implementation to DataWedgeHardwareScanner. Added HardwareScanner interface. Added ScannerController abstract class, for extendibility, added default implementation ParityFlowScannerController. ParityFlowScannerController helps manage the wished scanner status by checking current scanner status before emitting on the scanner itself if the scanner already has the wished status.
