+++ kotlin-data-wedge-lib/src/main/java/dk/gls/kdw/configuration/scanner/HardwareScanner.kt
@@ -1,0 +1,10 @@
+package dk.gls.kdw.configuration.scanner
+
+import dk.gls.kdw.model.scanner.ScannerOutput
+import kotlinx.coroutines.flow.Flow
+
+interface HardwareScanner {
+    fun scannerOutputFlow(): Flow<ScannerOutput>
+    fun suspendScanner()
+    fun resumeScanner()
+}
